// src/models/task.ts
/** @deprecated */
export type Task = {
  /** @deprecated */
  day: string;
  /** @deprecated */
  customersId: number;
  /** @deprecated */
  customersName: string;
  /** @deprecated */
  projectsId: number | 0;
  /** @deprecated */
  projectsName: string | null;
  /** @deprecated */
  servicesId: number | null;
  /** @deprecated */
  servicesName: string | null;
  /** @deprecated */
  lumpsumsId: number | null;
  /** @deprecated */
  lumpsumsAmount: number | null;
  /** @deprecated */
  lumpsumsName: string | null;
  /** @deprecated */
  lumpsumsPrice: number | null;
  /** @deprecated */
  lumpsumsUnit: string | null;
  /** @deprecated */
  billable: 0 | 1;
  /** @deprecated */
  textsId: number | 0;
  /** @deprecated */
  text: string | null;
  /** @deprecated */
  timeSince: string;
  /** @deprecated */
  timeUntil: string;
  /** @deprecated */
  duration: number;
  /** @deprecated */
  durationTime: string;
  /** @deprecated */
  durationText: string;
  /** @deprecated */
  isClocking: boolean;
  /** @deprecated */
  hasJustLumpsums: boolean;
  /** @deprecated */
  revenue?: number;
};